requirements.txt
absl-py==0.15.0
astetik==1.13
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
chances==0.1.9
channelnorm-cuda==0.0.0
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
correlation-cuda==0.0.0
cycler==0.11.0
Cython==0.29.30
dataclasses==0.8
decorator==4.4.2
efficientnet==1.0.0
filelock==3.4.1
flatbuffers==1.12
gast==0.4.0
gensim==3.8.3
geonamescache==1.3.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.1.0
idna==3.3
imageio==2.13.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
joblib==1.1.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kerasplotlib==1.0
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
natsort==8.2.0
networkx==2.5.1
nlg-eval @ git+https://github.com/Maluuba/nlg-eval.git@7f7993035a2f4729a15d20040fd904933ea58767
nltk==3.4.5
numpy==1.19.5
nvidia-cublas==11.2.1.74
nvidia-cuda-cupti==11.1.69
nvidia-cuda-nvcc==11.1.74
nvidia-cuda-nvrtc==11.1.105
nvidia-cuda-runtime==11.1.74
nvidia-cudnn==8.0.4.30
nvidia-cufft==10.3.0.74
nvidia-curand==10.2.2.74
nvidia-cusolver==11.0.0.74
nvidia-cusparse==11.2.0.275
nvidia-dali-cuda110==0.27.0
nvidia-dali-nvtf-plugin==0.27.0+nv20.11
nvidia-nccl==2.8.2
nvidia-pyindex==1.0.9
nvidia-tensorboard==1.15.0+nv20.11
nvidia-tensorflow==1.15.4+nv20.11
nvidia-tensorrt==7.2.1.6
oauthlib==3.2.0
opencv-python==4.5.4.60
opt-einsum==3.3.0
pandas==1.1.5
parso==0.7.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.29
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.2.2
Pygments==2.12.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
resample2d-cuda==0.0.0
rsa==4.8
s3transfer==0.5.2
sacremoses==0.0.53
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sentencepiece==0.1.96
six==1.15.0
sklearn==0.0
smart-open==6.0.0
statsmodels==0.12.2
talos==1.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
Theano==1.0.5
threadpoolctl==3.0.0
tifffile==2020.9.3
tokenizers==0.5.0
torch==1.10.0
tqdm==4.64.0
traitlets==4.3.3
transformers==2.5.0
typing-extensions==3.7.4.3
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrangle==0.7.2
wrapt==1.12.1
xdg==5.1.1
zipp==3.6.0
